// src/types/queryTypes.ts
export interface QueryRow {
  Campus:               string;
  Session:              string;
  Location:             string;
  Subject:              string;
  Catalog:              number | string;
  Descr:                string;
  Section:              number | string;
  Assoc:                number;
  "Class Nbr":          number;
  "Start Date":         string;
  "End Date":           string;
  "Facil ID":           string;
  Mon:                  string;
  Tue:                  string;
  Wed:                  string;
  Thu:                  string;
  Fri:                  string;
  Sat:                  string;
  Sun:                  string;
  Time_Start:           string;
  Time_End:             string;
  Mode:                 string;
  Consent:              string;
  "Class Stat":         string;
  "Tot Enrl":           number;
  Cap:                  number;
  ID:                   number | string;
  Name:                 string;
  "Note Nbr":           number | string;
  "Print Note":         string;
  Descrlong_A_Clsnotes: string;
  "Class Fee":          number | string;
  "Crse Fee":           number | string;
  "Sch Print":          string;
  "Crse Attr":          string;
  "File Type":          string;
  "Acad Org":           number | string;
  Acad_Org_Descr:       string;
  Term:                 number;
  Date_Time:            string;
};